WIP:  Open up the Attr class hierarchy for extension
This adds a new "InternalPluginBaseAttr" attribute node that can be
attached to AST nodes.

This one uses RTTIExtends things to allows plugins to inherit from it.

This allows plugins to attach things that are not "AnnotationAttr" to
ast nodes.

When does it need to delegate to a virtual method in plugin instead of
the switching over kind?